backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
anthropic==0.28.0
pydantic==2.7.1
pydantic-settings==2.2.1
supabase==2.4.2
redis==5.0.4
celery==5.4.0
python-multipart==0.0.9
sse-starlette==2.1.0
reportlab==4.2.0

# ── CrewAI Multi-Agent System ─────────────────────────────────────────
crewai[tools]>=0.177.0,<1.0.0
semanticscholar>=0.8.4
